fix logout issue refactor viewmodels
