+++ app/build.gradle.kts
@@ -90,5 +90,4 @@ dependencies {
     implementation(libs.androidx.core.splashscreen)
     implementation(libs.otpverification)
 
-
 }
