+++ app/src/main/java/com/mobile/xplore_nu/SplashViewModel.kt
@@ -1,0 +1,22 @@
+package com.mobile.xplore_nu
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
+import com.mobile.domain.usecases.DataStoreUseCase
+import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.flow.SharingStarted
+import kotlinx.coroutines.flow.stateIn
+import javax.inject.Inject
+
+@HiltViewModel
+class SplashViewModel @Inject constructor(
+    dataStoreUseCase: DataStoreUseCase
+) : ViewModel() {
+
+    val isLoggedIn = dataStoreUseCase.getIsLoggedIn()
+        .stateIn(
+            scope = viewModelScope,
+            started = SharingStarted.WhileSubscribed(5_000),
+            initialValue = null,
+        )
+}
