+++ app/src/main/java/com/mobile/xplore_nu/ui/screens/auth/forgotPassword/PasswordResetPage.kt
@@ -1,4 +1,4 @@
-package com.mobile.xplore_nu.ui.screens.auth
+package com.mobile.xplore_nu.ui.screens.auth.forgotPassword
 
 import android.widget.Toast
 import androidx.compose.foundation.layout.Column
@@ -18,6 +18,7 @@ import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.focus.FocusDirection
 import androidx.compose.ui.focus.FocusManager
 import androidx.compose.ui.focus.FocusRequester
 import androidx.compose.ui.focus.focusRequester
@@ -31,7 +32,7 @@ import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.mobile.domain.models.RequestOtpResponse
+import com.mobile.domain.models.ResetPasswordResponse
 import com.mobile.domain.utils.Resource
 import com.mobile.domain.utils.Status
 import com.mobile.xplore_nu.ui.components.AppNameHeader
@@ -40,16 +41,17 @@ import com.mobile.xplore_nu.ui.components.LeftChevron
 import com.mobile.xplore_nu.ui.components.OutlinedTextFieldComponent
 import com.mobile.xplore_nu.ui.components.RedButton
 import com.mobile.xplore_nu.ui.theme.fontFamily
-import com.mobile.xplore_nu.ui.uistates.ForgotPasswordState
+import com.mobile.xplore_nu.ui.uistates.ResetPasswordState
 
 @Composable
-fun ForgotPasswordPage(
-    forgotPasswordState: ForgotPasswordState,
+fun PasswordResetPage(
+    resetPasswordState: ResetPasswordState,
     onBackButtonClicked: () -> Unit,
-    onEmailUpdated: (email: String) -> Unit,
-    onRequestOtpClicked: (email: String) -> Unit,
-    requestOtpResponse: Resource<RequestOtpResponse>,
-    navigateToVerifyOtpScreen: (email: String) -> Unit
+    onResetButtonClicked: () -> Unit,
+    navigateToResetConfirmationPage: () -> Unit,
+    resetPasswordResponse: Resource<ResetPasswordResponse>,
+    onPasswordUpdated: (String) -> Unit,
+    onConfirmPasswordUpdated: (String) -> Unit,
 ) {
     val focusRequester = remember {
         FocusRequester()
@@ -63,14 +65,14 @@ fun ForgotPasswordPage(
 
     val context = LocalContext.current
 
-    LaunchedEffect(requestOtpResponse) {
-        when (requestOtpResponse.status) {
+    LaunchedEffect(resetPasswordResponse) {
+        when (resetPasswordResponse.status) {
             Status.SUCCESS -> {
-                navigateToVerifyOtpScreen(forgotPasswordState.email)
+                navigateToResetConfirmationPage()
             }
 
             Status.ERROR -> {
-                Toast.makeText(context, requestOtpResponse.message, Toast.LENGTH_LONG).show()
+                Toast.makeText(context, resetPasswordResponse.message, Toast.LENGTH_LONG).show()
             }
 
             Status.LOADING -> {
@@ -98,7 +100,7 @@ fun ForgotPasswordPage(
         AppNameHeader()
         HuskyLogoImage()
         Text(
-            text = "Enter email to receive OTP",
+            text = "Reset Password",
             style = TextStyle(
                 fontSize = 24.sp,
                 fontFamily = fontFamily,
@@ -112,51 +114,76 @@ fun ForgotPasswordPage(
                 .fillMaxWidth()
                 .padding(start = 24.dp, end = 24.dp, bottom = 12.dp)
                 .focusRequester(focusRequester),
-            label = "Email ID",
-            value = forgotPasswordState.email,
-            onValueChange = onEmailUpdated,
-            isError = forgotPasswordState.email.isNotEmpty() && !forgotPasswordState.isEmailValid,
-            errorMessage = "Enter a valid email ID",
+            label = "Password",
+            value = resetPasswordState.password,
+            onValueChange = onPasswordUpdated,
+            isError = resetPasswordState.password.isNotEmpty() && !resetPasswordState.isPasswordValid,
+            errorMessage = "Enter a valid password",
+            keyboardOptions = KeyboardOptions(
+                imeAction = ImeAction.Next,
+                keyboardType = KeyboardType.Password
+            ),
+            keyboardActions = KeyboardActions(
+                onNext = {
+                    focusManager.moveFocus(FocusDirection.Down)
+                }
+            ),
+            enabled = !resetPasswordState.isLoading
+        )
+        Spacer(modifier = Modifier.height(4.dp))
+        OutlinedTextFieldComponent(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(start = 24.dp, end = 24.dp, bottom = 12.dp),
+            label = "Confirm Password",
+            value = resetPasswordState.confirmPassword,
+            onValueChange = onConfirmPasswordUpdated,
+            isError = resetPasswordState.confirmPassword.isNotEmpty() && !resetPasswordState.isConfirmPasswordValid,
+            errorMessage = "Enter a valid confirm password",
             keyboardOptions = KeyboardOptions(
                 imeAction = ImeAction.Done,
-                keyboardType = KeyboardType.Email
+                keyboardType = KeyboardType.Password
             ),
             keyboardActions = KeyboardActions(
                 onDone = {
                     focusManager.clearFocus()
                 }
             ),
-            enabled = !forgotPasswordState.isLoading
+            enabled = !resetPasswordState.isLoading
         )
-        Spacer(modifier = Modifier.height(24.dp))
+        Spacer(modifier = Modifier.height(64.dp))
         RedButton(
-            label = "Get OTP", modifier = Modifier
+            label = "Done ",
+            modifier = Modifier
                 .fillMaxWidth()
                 .height(54.dp)
                 .padding(start = 24.dp, end = 24.dp),
             onClick = {
-                onRequestOtpClicked(forgotPasswordState.email)
+                onResetButtonClicked()
             },
-            enabled = forgotPasswordState.canRequestOtp,
-            isLoading = forgotPasswordState.isLoading
+            enabled = resetPasswordState.isResetButtonEnabled,
+            isLoading = resetPasswordState.isLoading
         )
     }
 }
 
-@Preview(showSystemUi = true, showBackground = true)
+@Preview(
+    showBackground = true,
+    showSystemUi = true
+)
 @Composable
-fun ForgotPasswordPreview() {
-    ForgotPasswordPage(
-        forgotPasswordState = ForgotPasswordState(
-            canRequestOtp = true,
-            isLoading = false
-        ),
+fun PasswordResetPagePreview(
+) {
+    PasswordResetPage(
+        resetPasswordState = ResetPasswordState(),
         onBackButtonClicked = {},
-        onEmailUpdated = {},
-        onRequestOtpClicked = {},
-        navigateToVerifyOtpScreen = {},
-        requestOtpResponse = Resource.success<RequestOtpResponse>(
-            RequestOtpResponse("")
-        )
+        onResetButtonClicked = {},
+        navigateToResetConfirmationPage = {},
+        Resource.success(
+            ResetPasswordResponse("")
+        ),
+        onPasswordUpdated = {},
+        onConfirmPasswordUpdated = {}
     )
+
 }
