+++ app/src/main/java/com/mobile/xplore_nu/ui/screens/tour/TourPage.kt
@@ -4,13 +4,16 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
+import androidx.compose.material3.Button
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 
 @Composable
-fun TourPage(modifier: Modifier = Modifier) {
+fun TourPage(
+    onButtonClicked: () -> Unit
+) {
     Column(
         horizontalAlignment = Alignment.CenterHorizontally,
         modifier = Modifier
@@ -18,5 +21,8 @@ fun TourPage(modifier: Modifier = Modifier) {
             .verticalScroll(rememberScrollState())
     ) {
         Text("Home screen")
+        Button(onClick = onButtonClicked) {
+            Text(text = "Logout")
+        }
     }
 }
