+++ build.gradle.kts
@@ -1,5 +1,3 @@
-import org.jetbrains.kotlin.js.translate.context.Namer.kotlin
-
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 plugins {
     alias(libs.plugins.android.application) apply false
@@ -9,4 +7,5 @@ plugins {
     id("com.google.devtools.ksp") version "2.0.21-1.0.27" apply false
     alias(libs.plugins.compose.compiler) apply false
     alias(libs.plugins.hiltAndroid) apply false
+    id("com.google.protobuf") version "0.9.4" apply false
 }
