+++ data/src/main/java/com/mobile/data/local/mappers/AuthenticationResponseMapper.kt
@@ -1,0 +1,11 @@
+package com.mobile.data.local.mappers
+
+import com.mobile.data.remote.models.DAuthenticationResponse
+import com.mobile.domain.models.AuthenticationResponse
+
+fun DAuthenticationResponse.toAuthenticationResponse() = AuthenticationResponse(
+    message, token, dUser.toUser()
+)
+
+fun AuthenticationResponse.toDAuthenticationResponse() =
+    DAuthenticationResponse(message, token, user.toDUser())
