--- data/src/main/java/com/mobile/data/local/mappers/UserResponseMapper.kt
@@ -1,8 +1,0 @@
-package com.mobile.data.local.mappers
-
-import com.mobile.data.remote.models.DUserResponse
-import com.mobile.domain.models.UserResponse
-
-fun DUserResponse.toUserResponse() = UserResponse(id, firstName, lastName, email, role)
-
-fun UserResponse.toDUserResponse() = DUserResponse(id, firstName, lastName, email, role)
