+++ data/src/main/java/com/mobile/data/local/models/DUser.kt
@@ -1,8 +1,11 @@
 package com.mobile.data.local.models
 
+import android.os.Parcelable
 import androidx.room.Entity
 import androidx.room.PrimaryKey
+import kotlinx.parcelize.Parcelize
 
+@Parcelize
 @Entity
 data class DUser(
     @PrimaryKey
@@ -11,4 +14,4 @@ data class DUser(
     val lastName: String,
     val email: String,
     val role: String
-)
+) : Parcelable
