+++ data/src/main/java/com/mobile/data/remote/UserService.kt
@@ -1,16 +1,16 @@
 package com.mobile.data.remote
 
+import com.mobile.data.remote.models.DAuthenticationResponse
+import com.mobile.data.remote.models.DLoginRequest
+import com.mobile.data.remote.models.DLogoutResponse
 import com.mobile.data.remote.models.DRequestOtpRequest
 import com.mobile.data.remote.models.DRequestOtpResponse
-import com.mobile.data.remote.models.DLoginRequest
-import com.mobile.data.remote.models.DLoginResponse
 import com.mobile.data.remote.models.DResendOtpRequest
 import com.mobile.data.remote.models.DResendOtpResponse
 import com.mobile.data.remote.models.DResetPasswordRequest
 import com.mobile.data.remote.models.DResetPasswordResponse
-import com.mobile.data.remote.models.DUserResponse
 import com.mobile.data.remote.models.DUserRegisterBody
-import com.mobile.data.remote.models.DUserRegisterResponse
+import com.mobile.data.remote.models.DUserResponse
 import com.mobile.data.remote.models.DVerifyOtpRequest
 import com.mobile.data.remote.models.DVerifyOtpResponse
 import retrofit2.Response
@@ -21,28 +21,28 @@ import retrofit2.http.POST
 interface UserService {
 
     @POST("/auth/register")
-    suspend fun registerUser(@Body requestBody: DUserRegisterBody): Response<DUserRegisterResponse>
+    suspend fun registerUser(@Body requestBody: DUserRegisterBody): Response<DAuthenticationResponse>
 
     @POST("/auth/login")
-    suspend fun loginUser(@Body dLoginRequest: DLoginRequest): Response<DLoginResponse>
+    suspend fun loginUser(@Body dLoginRequest: DLoginRequest): Response<DAuthenticationResponse>
 
     @GET("/auth/Users")
     suspend fun getUsers(): Response<List<DUserResponse>>
 
     @POST("/auth/logout")
-    suspend fun logoutUser(): Response<DUserRegisterResponse>
+    suspend fun logoutUser(): Response<DLogoutResponse>
 
     @POST("/auth/forgot_password")
-    suspend fun requestOtp(@Body dRequestOtpRequest: DRequestOtpRequest ): Response<DRequestOtpResponse>
+    suspend fun requestOtp(@Body dRequestOtpRequest: DRequestOtpRequest): Response<DRequestOtpResponse>
 
     @POST("auth/verify_otp")
-    suspend fun verifyOtp(@Body dVerifyOtpRequest: DVerifyOtpRequest ): Response<DVerifyOtpResponse>
+    suspend fun verifyOtp(@Body dVerifyOtpRequest: DVerifyOtpRequest): Response<DVerifyOtpResponse>
 
     @POST("auth/resend_otp")
-    suspend fun resendOtp(@Body dResendOtpRequest: DResendOtpRequest ): Response<DResendOtpResponse>
+    suspend fun resendOtp(@Body dResendOtpRequest: DResendOtpRequest): Response<DResendOtpResponse>
 
     @POST("auth/reset_password")
-    suspend fun resetPassword(@Body dResetPasswordRequest: DResetPasswordRequest ): Response<DResetPasswordResponse>
+    suspend fun resetPassword(@Body dResetPasswordRequest: DResetPasswordRequest): Response<DResetPasswordResponse>
 
 
 }
