+++ data/src/main/java/com/mobile/data/remote/models/DAuthenticationResponse.kt
@@ -2,13 +2,14 @@ package com.mobile.data.remote.models
 
 import android.os.Parcelable
 import com.google.gson.annotations.SerializedName
+import com.mobile.data.local.models.DUser
 import kotlinx.parcelize.Parcelize
 
 @Parcelize
-data class DLoginResponse(
+data class DAuthenticationResponse(
     val message: String,
     val token: String,
     @SerializedName("user")
-    val dUser: DLoggedInUser
+    val dUser: DUser
 ) : Parcelable
