+++ data/src/main/java/com/mobile/data/remote/models/DLogoutResponse.kt
@@ -1,0 +1,6 @@
+package com.mobile.data.remote.models
+
+data class DLogoutResponse(
+    val message: String,
+)
+
