--- data/src/main/java/com/mobile/data/remote/models/DUserRegisterResponse.kt
@@ -1,7 +1,0 @@
-package com.mobile.data.remote.models
-
-data class DUserRegisterResponse(
-    val message: String,
-    val token: String,
-    val user: DLoggedInUser
-)
