+++ domain/build.gradle.kts
@@ -3,7 +3,6 @@ plugins {
     alias(libs.plugins.jetbrains.kotlin.jvm)
     id("com.google.devtools.ksp")
 }
-
 java {
     sourceCompatibility = JavaVersion.VERSION_17
     targetCompatibility = JavaVersion.VERSION_17
