--- domain/src/main/java/com/mobile/domain/models/LoggedInUser.kt
@@ -1,9 +1,0 @@
-package com.mobile.domain.models
-
-data class LoggedInUser(
-    val id: String,
-    val firstName: String,
-    val lastName: String,
-    val email: String,
-    val role: String
-)
