+++ domain/src/main/java/com/mobile/domain/models/LogoutResponse.kt
@@ -1,0 +1,5 @@
+package com.mobile.domain.models
+
+data class LogoutResponse (
+    val message: String
+)
