--- domain/src/main/java/com/mobile/domain/models/UserRegisterResponse.kt
@@ -1,7 +1,0 @@
-package com.mobile.domain.models
-
-data class UserRegisterResponse(
-    val message: String,
-    val token: String,
-    val user: LoggedInUser
-)
