+++ domain/src/main/java/com/mobile/domain/repository/UserRepository.kt
@@ -1,7 +1,8 @@
 package com.mobile.domain.repository
 
+import com.mobile.domain.models.AuthenticationResponse
 import com.mobile.domain.models.LoginRequest
-import com.mobile.domain.models.LoginResponse
+import com.mobile.domain.models.LogoutResponse
 import com.mobile.domain.models.RequestOtpRequest
 import com.mobile.domain.models.RequestOtpResponse
 import com.mobile.domain.models.ResendOtpRequest
@@ -10,7 +11,6 @@ import com.mobile.domain.models.ResetPasswordRequest
 import com.mobile.domain.models.ResetPasswordResponse
 import com.mobile.domain.models.User
 import com.mobile.domain.models.UserRegisterBody
-import com.mobile.domain.models.UserRegisterResponse
 import com.mobile.domain.models.VerifyOtpRequest
 import com.mobile.domain.models.VerifyOtpResponse
 import com.mobile.domain.utils.Resource
@@ -22,17 +22,19 @@ interface UserRepository {
 
     fun getUser(id: String): User?
 
-    suspend fun loginUser(loginRequest: LoginRequest): Resource<LoginResponse>
+    suspend fun loginUser(loginRequest: LoginRequest): Resource<AuthenticationResponse>
 
-    suspend fun registerUser(userRegisterBody: UserRegisterBody): Resource<UserRegisterResponse>
+    suspend fun registerUser(userRegisterBody: UserRegisterBody): Resource<AuthenticationResponse>
 
-    suspend fun logoutUser(): UserRegisterResponse
+    suspend fun logoutUser(): Resource<LogoutResponse>
 
     suspend fun setIsLoggedIn(isLoggedIn: Boolean)
 
     val isLoggedIn: Flow<Boolean>
 
     suspend fun clearAuthToken()
+
+    suspend fun clearIsLoggedIn()
 
     val authToken: Flow<String?>
